ColumnArray

+++ selfdrive/common/modeldata.h
@@ -5,7 +5,7 @@ const int LON_MPC_N = 32;
 const float MIN_DRAW_DISTANCE = 10.0;
 const float MAX_DRAW_DISTANCE = 100.0;
 
-const double T_IDXS[TRAJECTORY_SIZE] = {
+const float T_IDXS[TRAJECTORY_SIZE] = {
         0.        ,  0.00976562,  0.0390625 ,  0.08789062,  0.15625   ,
         0.24414062,  0.3515625 ,  0.47851562,  0.625     ,  0.79101562,
         0.9765625 ,  1.18164062,  1.40625   ,  1.65039062,  1.9140625 ,
@@ -13,7 +13,7 @@ const double T_IDXS[TRAJECTORY_SIZE] = {
         3.90625   ,  4.30664062,  4.7265625 ,  5.16601562,  5.625     ,
         6.10351562,  6.6015625 ,  7.11914062,  7.65625   ,  8.21289062,
         8.7890625 ,  9.38476562, 10.};
-const double X_IDXS[TRAJECTORY_SIZE] = {
+const float X_IDXS[TRAJECTORY_SIZE] = {
          0.    ,   0.1875,   0.75  ,   1.6875,   3.    ,   4.6875,
          6.75  ,   9.1875,  12.    ,  15.1875,  18.75  ,  22.6875,
         27.    ,  31.6875,  36.75  ,  42.1875,  48.    ,  54.1875,
